A2DP SRC: solve a silent issue at consecutive audio streams.
Rootcase: Bluetooth service may receive audio data for a while
after AVDTP suspended. Unfortunately the reading procedure is removed
from this point, therefore the last segment of audio may stay
in the data channel.
These data would become a start of the next audio stream.
A more severe issue occurs when the remaining data is not a
complete frame, which cause the audio sink failed to decode the
following audio frames.

Signed-off-by: Zihao Gao <gaozihao@xiaomi.com>