A2DP SRC: solve a silent issue at consecutive audio streams.
O62WT-4604

Rootcase: Bluetooth service may receive audio data for a while
after AVDTP suspended. Unfortunately the reading procedure is removed
from this point, therefore the last segment of audio may stay
in the data channel.
These data would become a start of the next audio stream.
A more severe issue occurs when the remaining data is not a
complete frame, which cause the audio sink failed to decode the
following audio frames.

Change-Id: I88ff6b0a16a4d7324736d51a796cc2f4241491a6
Signed-off-by: Zihao Gao <gaozihao@xiaomi.com>