Bluetooth: add STATE_SUSPENDING at A2DP stream.
O62WT-652

Rootcase: we used to have OFF/RUNNING/FLUSHING states to describe an audio stream.
which failed to distinguish:
1. the flushing period when Codec configured, and
2. the ending period when Audio is about to suspend.
We now add a new state SUSPENDING to indicate the ending period, in which all data
remaining in the audio socket is to be send out (later in another patch).
The mapping from the old states to new states are:
STATE_OFF -> STATE_STANDBY
STATE_FLUSHING -> STATE_FLUSHING or STATE_SUSPENDING
STATE_RUNNING -> STATE_STARTED

Change-Id: If3f4dae0974062802cdf7740832ade003f7d97ee
Signed-off-by: Zihao Gao <gaozihao@xiaomi.com>