A2DP: refine the A2DP suspend procedure at SRC.
O62WT-652

Rootcause: A2DP_CTRL_CMD_STOP cames earlier than all the buffered data
is send. The ending frames would lost.
To solved this issue, we added a STREAM_STOP_REQ on the first time that
audio is about to stop. Uppon this event handled, the audio stream would
switch from STARTED to SUSPENDING, and the buffered data can be send out
for a while.
If no more data can be acquired from the circle buffer, a
STREAM_SUSPEND_REQ will be triggered to stop the AVDTP stream.
The audio channel stops it's work when AVDTP suspend completed.

Change-Id: Ifb5f45909358f32379580aab10bbbcbb624bb261
Signed-off-by: Zihao Gao <gaozihao@xiaomi.com>