bttool:Implement bttool to automatically reply to pair request and confirm yes.
The vela device requires confirm yes or no, The bttool auto reply function is designed to confirm yes by default, which is not supported in the rel-4.0 feature. Therefore, this patch is submitted. However, if passkey entry or pin code input is required, auto reply will not work. You need to use the pincode or passkey command keyword of bttool pair to enter the content.

Signed-off-by: zhongzhijie1 <zhongzhijie1@xiaomi.com>