Bluetooth: Add a check for null pointer before reading or writing data in transport.
Rootcause: When the transport is closed, the pipe is set to null. If data is written before the transport's state changes to disconnected, it will result in accessing a null pointer and cause a crash. This issue is resolved by adding a null pointer check before reading or writing data.

Signed-off-by: jialu <jialu@xiaomi.com>