Bluetooth: Add a check for null pointer before reading or writing data in transport.
O65-16442

Rootcause: When the transport is closed, the pipe is set to null. If data is written before the transport's state changes to disconnected, it will result in accessing a null pointer and cause a crash. This issue is resolved by adding a null pointer check before reading or writing data.

Change-Id: I32d4e6a7b93b21c6d67d8932e7cc2ce3cbb6a638
Signed-off-by: jialu <jialu@xiaomi.com>