64-bit: Adjust the platform configuration macros in bt_config.h.
VELAPLATFO-59352

Rootcause: The configuration file of NUTTX is included after the platform macro configuration, which causes the inability to identify the compilation platform bit of Vela. To properly define the platform configuration macro, the following two adjustments are made: 1. Move the platform configuration macro to the end of bt_config.h; 2. Add compilation detection for the x86_64 platform.

Change-Id: If37a2d0352388da4169b336d44e638a44d156a0c
Signed-off-by: chejinxian1 <chejinxian1@xiaomi.com>