64-bit: Adjust the platform configuration macros in bt_config.h.
Rootcause: The configuration file of NUTTX is included after the platform macro configuration, which causes the inability to identify the compilation platform bit of Vela. To properly define the platform configuration macro, the following two adjustments are made: 1. Move the platform configuration macro to the end of bt_config.h; 2. Add compilation detection for the x86_64 platform.

Signed-off-by: chejinxian1 <chejinxian1@xiaomi.com>