SPP: Fix the issue of improperly closing the pipe.
VELAPLATFO-53158

Rootcause: When using SPP capability in cross-core scenario, it is necessary to perform cross-core communication through RPMSG. Since the Bluetooth service does not know which pipe connection method the application uses when establishing an SPP connection, it waits for the application to connect using both local and RPMSG pipes, and closes the unused pipe after connecting. When not in a cross-core scenario, the Bluetooth service does not use both pipes, which causes the subsequent closing behavior to result in runtime errors, causing the Bluetooth thread to be cleared by the system, thereby stopping the Bluetooth service.

Change-Id: Id650f91eb04d23d04dbf0203603431806c58d0b6
Signed-off-by: chejinxian1 <chejinxian1@xiaomi.com>