bluetooth: Fix hfp hf loop logic error causing skipped nodes during linked list traversal.
VELAPLATFO-62711

Rootcause:  suppose the linked list is A -> B -> C. When processing node A, if A is removed, tmp becomes B. At this point, cnode is assigned to B, and the loop continues. However, during the iteration step of the for loop, cnode is set to bt_list_next(clist, B) (which returns C). As a result, node B is skipped and never processed.

Change-Id: Ic91927eb46903c0cfc6c443d92bb79551fffc24a
Signed-off-by: zhangyuan20 <zhangyuan20@xiaomi.com>