zblue: Fix GATT reconnection failure after disconnection.
VELAPLATFO-60493

Root cause:
The Zephyr BLE stack automatically unreferences connection resources used internally, such as ATT and L2CAP.
However, when acting as a CENTRAL (initiator) and calling `bt_conn_le_create()`, an explicit reference to the connection is returned.
This reference is not automatically released by the stack, so it must be explicitly unref'd by the application.

Fix:
In the disconnection callback, call `bt_conn_unref()` for CENTRAL role connections to avoid reference leaks
and ensure that future reconnections are not blocked.

Change-Id: Ica2afb31efa9155f2bb568561c8109b7f68b5896
Signed-off-by: zhongzhijie1 <zhongzhijie1@xiaomi.com>