Adapter: Modify bond state notification method.
Rootcause: Modify the implementation of the bond state notification to support reporting the previous state. To more accurately report the bond state at the Framework layer, this modification will trigger a notification event when changing the bond state. To avoid long-term occupation of the adapter_lock, the method of do_in_service_loop is used to execute the actual notification.

Signed-off-by: chejinxian1 <chejinxian1@xiaomi.com>