bluetooth: core: release reference after using ACL conn.
VELAPLATFO-56422

rootcause: In the bt_conn_lookup_addr_br() function used by AVRCP and A2DP, a reference to a conn is obtained. After using the conn, the reference is not released, causing the reference to remain non-zero after the ACL is disconnected, leading to uncleaned resources. When the connection attempts exceed 5 times, connection failures may occur.

Change-Id: I2219bb6a55c05f256945c7732e0ff3f9a0096522
Signed-off-by: jialu <jialu@xiaomi.com>