Optimization of power consumption in re-connect when disconnected(0x08).
GTKGLASS-12123

Increase in re-connect interval from 8s to 12s. In the current power consumption test, the power consumption is 4.67% per half hour when the pullback interval is 8s, and 3.72% when it is increased to 12s. Moreover, the reconnection interval is too long, resulting in poor user experience. Considering the power consumption and user experience, 12s is finally chosen as the reconnection interval.

Change-Id: Ifd2484bc0783d8b07e397e18b7432dfd74e4da59
Signed-off-by: liuxiang18 <liuxiang18@xiaomi.com>