bluetooth:  Rename the macro for modifying A2DP SBC parameters.
bug: v/60871

Root cause: The macro names A2DP_SBC_SUBBAND_MSK and A2DP_SBC_SUBBAND_4 conflict with the macros defined in zblue, causing warnings to appear. The names of the macros in the framework were prefixed with BT to resolve the conflict, and the macro names for A2DP SBC defined in the framework were also uniformly prefixed with BT to maintain consistency in style.

Signed-off-by: jialu <jialu@xiaomi.com>