Android-Vela: add clip callback for hf
O61-4338

Rootcause: The application wants to get the name of the hf call by clip. The order is clip first, then clcc, but clip can only be called back after clcc, resulting in the application not being able to get the name. A separate callback needs to be added to clip.

Change-Id: I9713db812060ba8ca002460e819a7ef5229313e4
Signed-off-by: zhangyuan20 <zhangyuan20@xiaomi.com>