Bluetooth: Fix the issue of the reader in euv_pipe not being correctly released.
O61N-22663

Rootcause: When a pipe connection is successfully established and ready to read data, a reader will be allocated to handle the reading process, and the reader will be associated with the pipe object. When it is necessary to stop reading from the pipe, the reader needs to be released. Specifically, when a pipe disconnection occurs, the allocated reader also needs to be released. However, when a disconnection is initiated by either end of the pipe, the other end may read the EOF symbol from the reader first, causing the pipe to be set to a non-active state. This prevents the current logic of first checking the pipe status before releasing the reader in euv_pipe_read_stop from releasing the reader.

Change-Id: I8b80f04bce79f81ef6296b3dfd7d8307c13c210d
Signed-off-by: chejinxian1 <chejinxian1@xiaomi.com>