bluetooth: fix the problem of unsynchronized callback of connection address and bond address.
VELAPLATFO-62114

The current callback uses the info.le.dst as the callback address, but the dst addr is the random address of the peer when it is unbound. After bonded, the dst addr is updated to the public address after parsing. The result is that after bonded, the connected/disconnect callback is public address, which will result in first connected random address cannot to be removed.

Change-Id: I71a8ada5834c577e67503d01f979796d00085cd1
Signed-off-by: liuxiang18 <liuxiang18@xiaomi.com>