bluetooth: fix the problem of unsynchronized callback of connection address and bond address.
The current callback uses the info.le.dst as the callback address, but the dst addr is the random address of the peer when it is unbound. After bonded, the dst addr is updated to the public address after parsing. The result is that after bonded, the connected/disconnect callback is public address, which will result in first connected random address cannot to be removed.

Signed-off-by: liuxiang18 <liuxiang18@xiaomi.com>