SPP: Fix the potential memory leak after SPP disconnection.
O61-15429

Rootcause: When the SPP Proxy connection has not been successfully established, disconnection occurs, causing the cached data to be released by the protocol stack, and then the list managing the cached data cannot be released. Therefore, when releasing the software resources of the SPP connection, the cache management list needs to be released.

Change-Id: Ic5543309c81c4c4041eea42014731879a66f0a36
Signed-off-by: chejinxian1 <chejinxian1@xiaomi.com>