gatt: Fix incorrect MTU reporting by using the minimum of tx and rx
VELAPLATFO-66910

Problem:
When acting as a GATT server, if the local supported MTU is 498 and the client sends an Exchange MTU Request of 517,
we incorrectly report 517 to the upper layer in the MTU update callback.

This causes issues in throughput tests — tools like bttool attempt to send full 517-byte packets, which exceed the local limit.
As a result, the stack rejects the buffers and the data fails to reach the controller.

Fix:
Always report the minimum of tx and rx as the effective MTU to the upper layer, ensuring compatibility between both sides.

Change-Id: Ic04033dca83008d10cd86b425a77b0a954e25c60
Signed-off-by: zhongzhijie1 <zhongzhijie1@xiaomi.com>