bluetooth: fix bond fail(peer initiate pair) fail when peer devcice deletes bond Info.
VELAPLATFO-66770

The previous modification only addressed the removal of the key when pairing as the central device failed again. The processing logic for the device acting as both the central and peripheral devices is not consistent. When acting as the central device, if a key is detected during this session, the encryption process is initiated directly. The peer returns BT_SECURITY_ERR_PIN_OR_KEY_MISSING. When acting as the peripheral device, BT_SECURITY_ERR_AUTH_REQUIREMENT is reported.

Change-Id: I6b84b3b92381afe212ec93ef270e3bda4c5b20c0
Signed-off-by: liuxiang18 <liuxiang18@xiaomi.com>