bluetooth: Fix the issue of scanner state not being cleared due to closing Bluetooth during scanning, which leads to scan failure again.
VELAPLATFO-71516

Root cause: The state of the scanner is only cleared when the scan is stopped or closed. When Bluetooth is turned off while scanning is in progress, the state of the scanner is not cleared. This leads to the failure to start scanning when Bluetooth is turned back on.

Change-Id: Ib851e8e03bfbecf72797dd51fd926c56aee52832
Signed-off-by: jialu <jialu@xiaomi.com>