scanners initiated by multiple apps, select the optimal scan parameters and filter policy.
bug: v/54956

rootcause: Originally, multiple scanners were not supported, and the new scanner could only passively monitor the scanning results of the old scanner. Now, the global optimal scanner will be selected, and when a scan fails, all scanners will be notified that the current scan has stopped. And free scanner resources.

Signed-off-by: zhongzhijie1 <zhongzhijie1@xiaomi.com>