bluetooth: Fix the issue where AVRCP fails to retrieve song information.
VELAPLATFO-79788

When the length of a song's attribute is 0, both types and chr_sets of that attribute are set to the default value 0. During message processing on the client side, the attribute's text is not assigned a NULL value due to types = 0. This leads to an error during string length calculation (strlen) when the feature processes the callback.

Change-Id: I1b9f97375f2fc8ec12035170745a13663d13c95a
Signed-off-by: jialu <jialu@xiaomi.com>