bluetooth: Fix the issue where AVRCP fails to retrieve song information.
When the length of a song's attribute is 0, both types and chr_sets of that attribute are set to the default value 0. During message processing on the client side, the attribute's text is not assigned a NULL value due to types = 0. This leads to an error during string length calculation (strlen) when the feature processes the callback.

Signed-off-by: jialu <jialu@xiaomi.com>